CBG-5: Split graphql data classes into model dir
