+++ Backend/BlameSightBackend/BlameSightBackend/Services/GraphqlService.cs
@@ -60,121 +60,3 @@ public class GraphQLClient
         }
     }
 }
-
-public class GraphField
-{
-    string name;
-    Attributes? attributes;
-    GraphObject? graphObject;
-
-    public GraphField(string name, Attributes? attributes = null, GraphObject? graphObject = null)
-    {
-        this.name = name;
-        this.attributes = attributes;
-        this.graphObject = graphObject;
-    }
-
-    public override string ToString()
-    {
-        string result = name;
-        result += attributes == null ? "" : $"({attributes})";
-        result += graphObject == null ? "" : $"{{ {graphObject} }}";
-
-        return result;
-    }
-
-}
-
-public class GraphObject
-{
-    List<GraphField> fields;
-
-    public GraphObject(List<GraphField> list)
-    {
-        this.fields = list;
-    }
-
-    public override string ToString()
-    {
-        string result = string.Empty;
-        foreach (GraphField field in fields)
-        {
-            result += $"{field} \n";
-        }
-        return result;
-    }
-
-}
-
-public class GraphQuery
-{
-    public List<GraphField>? fields;
-
-    public GraphQuery() { }
-
-    public void add(GraphField field)
-    {
-        if (fields == null)
-        {
-            fields = [field];
-        }
-        else
-        {
-            fields.Add(field);
-        }
-    }
-
-
-
-    public override string ToString()
-    {
-        string result = "";
-
-        if (fields != null)
-        {
-            foreach (GraphField item in fields)
-            {
-                result += $"{item} \n";
-            }
-        }
-        string json = JsonSerializer.Serialize($"{{ {result} }}");
-        return $"{{ \"query\" :  {json} }}";
-    }
-}
-
-
-public class Attributes
-{
-    List<KeyValuePair<string, string>>? keyValuePairs;
-
-    public Attributes() { }
-    public Attributes(List<KeyValuePair<String, String>> keyValuePairs)
-    {
-        this.keyValuePairs = keyValuePairs;
-    }
-
-    public void add(string filtername, string value)
-    {
-        KeyValuePair<string, string> temp = new(filtername, value);
-        if (keyValuePairs != null)
-        {
-            keyValuePairs.Add(temp);
-        }
-        else
-        {
-            keyValuePairs = new List<KeyValuePair<string, string>>([temp]);
-        }
-    }
-    public override string ToString()
-    {
-        string result = "";
-        if (keyValuePairs != null)
-        {
-            foreach (KeyValuePair<string, string> item in keyValuePairs)
-            {
-                result += $"{item.Key}:\"{item.Value}\",";
-            }
-        }
-        return result.Remove(result.Length - 1);
-    }
-}
